Added elm-package.json reference
I was working through the tutorial and this through me for a bit of a loop.  My elm-package.json wasn't in the src directory so the compiler didn't immediately find modules.  Was only a few minutes of googling to figure out, but thought it would be nice to have in the tutorial.  

Also fixed some minor grammar/english mistakes.